Work on deleteToDo functionality
The Problem: When trying to delete the first to do after the initial page
load, it would also delete the second to do. It somehow calls the function
twice.
The Solution: Build the logic for deleting to dos directly into the
displayToDos(), this works, but does not follow the
SOLID design principles.
If the reason for the issue is understood better, this may be
improved.