Edit pen pug, code removal
When the edit pen was pressed on a task and than the modal was closed
without making changes and again the same edit button was pressed there
was an error thrown.
I removed the code which caused the error, but the functionality which would
be wishful is removed with it. The functionality should be:
- The user behavior from above should not throw an error, but should
    work as intended.
- When the edit button on a task is pressed and the task is edited it should
    remove the old edited task and add the new edited task when the add
    to do button in the form modal is pressed.